pyproject.toml
[tool.poetry]
name = "salud-publica-digital"
version = "0.1.0"
description = "Proyecto de Salud Pública Digital"
authors = ["Alejandro Dinamarca <a.dinamarca@canalcero.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
django = "^5.0.6"
django-crispy-forms = "^2.2"
crispy-bootstrap5 = "^2024.2"
pymongo = "^4.10.1"
djangorestframework = "^3.15.2"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
